requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
azure-cognitiveservices-vision-face==0.6.0
azure-common==1.1.28
azure-core==1.30.1
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
idna==3.7
importlib_resources==6.4.0
isodate==0.6.1
kiwisolver==1.4.5
matplotlib==3.8.4
msrest==0.7.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==3.4.17.61
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
typing_extensions==4.11.0
urllib3==1.26.15
zipp==3.18.1
